requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
flask-restx==1.3.0
psycopg2-binary==2.9.10
pgvector==0.4.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
gunicorn==21.2.0 